requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
channels==3.0.4
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
daphne==3.0.2
Django==5.1.2
djangorestframework==3.15.2
Flask==3.0.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
python-decouple==3.8
requests==2.32.3
service-identity==24.2.0
simplejson==3.19.3
sqlparse==0.5.3
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
zope.interface==7.2
